Validate if required claims are available on a Resource Link message
The required claims are:
 - LTI Version
 - Deployment ID
 - Target Link URI
 - Resource Link ID

There are also User ID and User Role claims that are stated to be
mandatory by the spec, but they also say that the platform may choose to
execute an anonymous launch, rendering those parameters optional. So all
in all -- the parameters are anonymous from our perspective (at least
for now).

[#171331378]