Factor a ResourceLinkService out
Potentially we'd like to move the method related to LTI 1.3
authentication response endpoint to a separate file since it's not
actually a resource someone would be interested in. I've added it there
for now to remove the direct usage of UriComponentsBuilder from different
places in the code and get on with it.

[#172888961]