Remove target="_parent" from redirects
ITWorx have reached out and say that this interrupts the workflow they
have designed. IMS spec is unclear about how this behaviour should work
and our choice of adding the _parent target was a little arbitrary and
not tested properly with any other integration partners, so I'm happy to
remove it for now.

We can make this later configurable as a launch parameter if needed.