WIP
Things seem to be working, there's a few marginal differences between
padding on the card, but maybe 1/2pxs. For some reason in
search-and-embed view it has a different background colour to that of
that in staging. And for some reason the little info buttons on both
pages weren't appearing (see the demo for what it should look like). Not
sure if that's just a setup thing.

We weren't defining a global font, so that is probably something we can tidy up a
little after this.