Provide separate view for responsive embed view
When we load our LTI screens, we pass some additional data to them.
Search&Embed view expects to get `deep_link_return_url` and
`deployment_id`, while Search expects to get `embeddable_video_url` and
`show_copy_link`.
To be able to showcase the new responsive views and components it's
easier if we follow the current setup and have separate views for these
instead of trying to mush the needed data together somehow.

[#177834609]